splitting the code
