+++ apps/client/app/Game/Board.js
@@ -1,265 +1,9 @@
 import React, { useState, useEffect, useRef } from "react";
 import Matter from "matter-js";
+import {CreatRackets, CreateBallFillWall} from "./Bodies"
+import ListenKey from "./Keys"
+import Collision from "./Collision"
 
-const CreatRackets = (Bodies, RacketWidth, RacketHeight, render) => {
-  // create the left Racket
-  const RacketLeft = Bodies.rectangle(13, 39, RacketWidth, RacketHeight, {
-    label: "RacketR",
-    isStatic: true,
-    restitution: 1,
-    friction: 0,
-    frictionAir: 0,
-    frictionStatic: 0,
-    render: {
-      fillStyle: "#EEEEEE",
-      lineWidth: 1,
-    },
-  });
-  // create the right Racket
-  const RacketRight = Bodies.rectangle(
-    render.options.width - 13,
-    render.options.height - 39,
-    RacketWidth,
-    RacketHeight,
-    {
-      label: "RacketL",
-      isStatic: true,
-      restitution: 1,
-      friction: 0,
-      frictionAir: 0,
-      frictionStatic: 0,
-      render: {
-        fillStyle: "#FFD369",
-        lineWidth: 1,
-      },
-    }
-  );
-
-  return { RacketLeft, RacketRight };
-};
-
-const CreateBallFillWall = (
-  Bodies,
-  render,
-  initialBallPos,
-  ignored,
-  engine
-) => {
-  //draw the Ball
-  const Ball = Bodies.circle(initialBallPos.x, initialBallPos.y, 20, {
-    restitution: 1,
-    friction: 0,
-    frictionAir: 0,
-    frictionStatic: 0,
-    inertia: Infinity,
-    render: {
-      lineWidth: 2,
-      fillStyle: "#00FFD1",
-    },
-    collisionFilter: {
-      mask: ~ignored,
-    },
-  });
-  //draw the Fil
-  const Fil = Bodies.rectangle(
-    render.canvas.width / 2,
-    render.canvas.height / 2,
-    1,
-    render.canvas.height,
-    {
-      isStatic: true,
-      collisionFilter: {
-        category: ignored,
-      },
-      render: {
-        fillStyle: "black",
-      },
-    }
-  );
-  //draw Walls
-  const Walls = [
-    Bodies.rectangle(render.canvas.width / 2, 0, render.canvas.width, 5, {
-      isStatic: true,
-      restitution: 1,
-      friction: 0,
-      frictionAir: 0,
-      frictionStatic: 0,
-      label: "top",
-      render: {
-        fillStyle: "#FFD369",
-      },
-    }),
-    Bodies.rectangle(
-      render.canvas.width / 2,
-      render.canvas.height,
-      render.canvas.width,
-      5,
-      {
-        isStatic: true,
-        restitution: 1,
-        friction: 0,
-        frictionAir: 0,
-        frictionStatic: 0,
-        label: "bottom",
-        render: {
-          fillStyle: "#FFD369",
-        },
-      }
-    ),
-    Bodies.rectangle(0, render.canvas.height / 2, 5, render.canvas.height, {
-      isStatic: true,
-      restitution: 1,
-      friction: 0,
-      frictionAir: 0,
-      frictionStatic: 0,
-      label: "left",
-      render: {
-        fillStyle: "#FFD369",
-      },
-    }),
-    Bodies.rectangle(
-      render.canvas.width,
-      render.canvas.height / 2,
-      5,
-      render.canvas.height,
-      {
-        isStatic: true,
-        restitution: 1,
-        friction: 0,
-        frictionAir: 0,
-        frictionStatic: 0,
-        label: "right",
-        render: {
-          fillStyle: "#FFD369",
-        },
-      }
-    ),
-  ];
-
-  return { Ball, Fil, Walls };
-};
-
-const ListenKey = (
-  render,
-  RacketRight,
-  RacketLeft,
-  Ball,
-  RacketHeight,
-  Body,
-  ballSpeed,
-  isStart,
-  setIsStart
-) => {
-  let keys = {};
-
-window.addEventListener("keydown", (event) => {
-	keys[event.code] = true;
-  
-	if (event.code === "Space") {
-	  setIsStart((prevIsStart) => {
-		if (prevIsStart) {
-		  Body.setVelocity(Ball, { x: 12, y: ballSpeed });
-		  console.log("Game started");
-		} else {
-		  console.log("Game already started");
-		}
-		return false; // Ensure that `isStart` is set to `false`
-	  });
-	}
-  });
-  
-  window.addEventListener("keyup", (event) => {
-    keys[event.code] = false;
-  });
-
-  //run each key separate
-  function RunMovement() {
-    let racketSpeed = 12;
-    const canvasHeight = render.options.height;
-    let drY = 0;
-    let dlY = 0;
-
-    if (keys["ArrowUp"]) {
-      drY -= racketSpeed;
-    }
-    if (keys["ArrowDown"]) {
-      drY += racketSpeed;
-    }
-
-    if (
-      RacketRight.position.y - RacketHeight / 2 + drY > 0 &&
-      RacketRight.position.y + RacketHeight / 2 + drY < canvasHeight
-    ) {
-      Matter.Body.translate(RacketRight, { x: 0, y: drY });
-    }
-
-    if (keys["KeyW"]) {
-      dlY -= racketSpeed;
-    }
-    if (keys["KeyS"]) {
-      dlY += racketSpeed;
-    }
-
-    if (
-      RacketLeft.position.y - RacketHeight / 2 + dlY > 0 &&
-      RacketLeft.position.y + RacketHeight / 2 + dlY < canvasHeight
-    ) {
-      Matter.Body.translate(RacketLeft, { x: 0, y: dlY });
-    }
-    //method that helps the browser to draw the object and to run smoothly
-    requestAnimationFrame(RunMovement);
-  }
-  RunMovement();
-  return () => {
-    window.removeEventListener("keydown", handleKeyDown);
-    window.removeEventListener("keyup", handleKeyUp);
-  };
-};
-
-const Collision = (
-  Events,
-  Body,
-  engine,
-  Ball,
-  setScoreA,
-  setScoreB,
-  initialBallPos,
-  setIsStart
-) => {
-  //sounds of the ball hitting the Racket and the wall
-  const BallSound = new Audio("./BallSound.mp3");
-  const Fail = new Audio("./Fail.mp3");
-
-  Events.on(engine, "collisionStart", (event) => {
-    const pairs = event.pairs;
-    pairs.forEach((pair) => {
-      const { bodyA, bodyB } = pair;
-      const ball = Ball;
-      let bodyC = 0;
-      //identify the other object that the ball will hit
-      if (bodyA === ball) bodyC = bodyB;
-      else if (bodyB === ball) bodyC = bodyA;
-
-      //apply sound and score depends on the other object
-      if (bodyC.label === "left") {
-        setScoreA((prevNumber) => prevNumber + 1);
-        setIsStart(true);
-        console.log("game ended");
-        Fail.play();
-        Body.setVelocity(Ball, { x: 0, y: 0 });
-        Body.setPosition(Ball, initialBallPos);
-      } else if (bodyC.label === "right") {
-        setScoreB((prevNumber) => prevNumber + 1);
-        setIsStart(true);
-        console.log("game ended");
-        Fail.play();
-        Body.setVelocity(Ball, { x: 0, y: 0 });
-        Body.setPosition(Ball, initialBallPos);
-      } else if (bodyC.label === "RacketR" || bodyC.label === "RacketL")
-        BallSound.play();
-    });
-  });
-};
 
 export function Game() {
   //initializing the canva and box
